refactor fetching node major varsions on integration.json
