+++ integration.json
@@ -3,5 +3,8 @@
   "ubi-nodejs-extension": "github.com/paketo-community/ubi-nodejs-extension",
   "node-engine": "github.com/paketo-buildpacks/node-engine",
   "npm-install": "github.com/paketo-buildpacks/npm-install",
-  "go-dist": "github.com/paketo-buildpacks/go-dist.git"
+  "go-dist": "github.com/paketo-buildpacks/go-dist",
+  "nodejs-major-versions": [
+    16, 18, 20
+  ]
 }
